+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## feature/bgp
-- Added support for RouterBgpNbrAF class.
+- Added support for RouterBgpNeighborAF class.
 
 ## [1.0.0] - 2015-08-28
 ### Added